+++ src/pages/Dashboard/components/PeopleList.jsx
@@ -1,12 +1,10 @@
 import PeopleListItem from './PeopleListItem'
-
 function PeopleList(props) {
   const { people } = props
-
   return (
     <ul>
       {people.map((person, index) => (
-        <PeopleListItem key={index} person={person} />
+            <PeopleListItem key={index} person={person}/>
       ))}
     </ul>
   )
